requirements.txt
phidata>=2.7.0
groq>=0.9.0
openai>=1.40.0
duckduckgo-search>=6.2.0
streamlit>=1.38.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
